+++ src/Lykke.Job.TransactionHandler/Commands/DestroyCommand.cs
@@ -1,0 +1,10 @@
+﻿using ProtoBuf;
+
+namespace Lykke.Job.TransactionHandler.Commands
+{
+    [ProtoContract]
+    public class DestroyCommand
+    {
+
+    }
+}
