+++ src/Lykke.Job.TransactionHandler/Commands/ExternalCashInCommand.cs
@@ -1,0 +1,12 @@
+﻿using ProtoBuf;
+
+namespace Lykke.Job.TransactionHandler.Commands
+{
+    [ProtoContract]
+    public class ExternalCashInCommand
+    {
+        public string ClientId { get; set; }
+        public string AssetId { get; set; }
+        public double Amount { get; set; }
+    }
+}
