+++ src/Lykke.Job.TransactionHandler/Commands/IssueCommand.cs
@@ -1,0 +1,9 @@
+﻿using ProtoBuf;
+
+namespace Lykke.Job.TransactionHandler.Commands
+{
+    [ProtoContract]
+    public class IssueCommand
+    {
+    }
+}
