+++ src/Lykke.Job.TransactionHandler/Commands/ManualUpdateCommand.cs
@@ -1,0 +1,9 @@
+﻿using ProtoBuf;
+
+namespace Lykke.Job.TransactionHandler.Commands
+{
+    [ProtoContract]
+    public class ManualUpdateCommand
+    {
+    }
+}
