+++ src/Lykke.Job.TransactionHandler/Queues/EthereumEventsQueue.cs
@@ -23,7 +23,6 @@ using Newtonsoft.Json.Converters;
 using Lykke.Job.TransactionHandler.Core.Services.BitCoin;
 using Lykke.Service.OperationsRepository.Client.Abstractions.CashOperations;
 using Lykke.Service.OperationsRepository.AutorestClient.Models;
-using Lykke.Job.TransactionHandler.Core.Services.Ethereum;
 
 namespace Lykke.Job.TransactionHandler.Queues
 {
