+++ src/Lykke.Job.TransactionHandler/Queues/IQueueSubscriber.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Lykke.Job.TransactionHandler.Queues
 {
