--- src/Lykke.Job.TransactionHandler/Queues/Messaging/CashInOutSaga.cs
@@ -1,7 +1,0 @@
-﻿namespace Lykke.Job.TransactionHandler.Queues.Messaging
-{
-    public class CashInOutSaga
-    {
-        
-    }
-}
