+++ src/Lykke.Job.TransactionHandler/Sagas/CashInOutSaga.cs
@@ -1,0 +1,7 @@
+﻿namespace Lykke.Job.TransactionHandler.Sagas
+{
+    public class CashInOutSaga
+    {
+        
+    }
+}
